requirements.txt
streamlit==1.28.1
langgraph==0.0.62
langchain==0.1.4
langchain-google-genai==0.0.8
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0
sqlite3
typing-extensions==4.8.0
pydantic==2.5.2
requests==2.31.0
beautifulsoup4==4.12.2